refactor version and add version command to sget
