+++ Makefile
@@ -39,7 +39,7 @@ ifeq ($(DIFF), 1)
     GIT_TREESTATE = "dirty"
 endif
 
-PKG=github.com/sigstore/cosign/cmd/cosign/cli/options
+PKG=github.com/sigstore/cosign/pkg/version
 
 LDFLAGS="-X $(PKG).GitVersion=$(GIT_VERSION) -X $(PKG).gitCommit=$(GIT_HASH) -X $(PKG).gitTreeState=$(GIT_TREESTATE) -X $(PKG).buildDate=$(BUILD_DATE)"
 
@@ -142,4 +142,4 @@ cosigned: lint ## Build cosigned binary
 
 .PHONY: sget
 sget: ## Build sget binary
-	go build -trimpath -o $@ ./cmd/sget
+	go build -trimpath -ldflags $(LDFLAGS) -o $@ ./cmd/sget
