+++ cmd/cosign/cli/options/registry.go
@@ -25,6 +25,8 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 	ociremote "github.com/sigstore/cosign/pkg/oci/remote"
 	"github.com/spf13/cobra"
+
+	"github.com/sigstore/cosign/pkg/version"
 )
 
 // RegistryOptions is the wrapper for the registry options.
@@ -65,7 +67,7 @@ func (o *RegistryOptions) ClientOpts(ctx context.Context) ([]ociremote.Option, e
 func (o *RegistryOptions) GetRegistryClientOpts(ctx context.Context) []remote.Option {
 	opts := []remote.Option{
 		remote.WithContext(ctx),
-		remote.WithUserAgent("cosign/" + VersionInfo().GitVersion),
+		remote.WithUserAgent("cosign/" + version.GetVersionInfo().GitVersion),
 	}
 
 	if o.KubernetesKeychain {
